testprj/requirements.txt
Django==1.10.4
djangorestframework==3.5.3
django-rest-swagger==0.3.10
python-dateutil==2.6.0
